+++ tests/integration/test_data/open3e_device_config/Open3Edatapoints_680.py
@@ -3,23 +3,22 @@ from open3e.Open3Ecodecs import *
 
 dataIdentifiers = {
     "name": "HPMUMASTER",
-    "dids" :
-    {
-        256 : None,
-        262 : None,
-        268 : None,
-        381 : None,
-        382 : None,
-        396 : None,
-        505 : None,
-        507 : None,
-        548 : None,
-        592 : None,
-        604 : None,
-        691 : None,
-        692 : None,
-        1006 : None,
-        1007 : None,
-        1097 : None,
-    }
+    "dids": {
+        256: None,
+        262: None,
+        268: None,
+        381: None,
+        382: None,
+        396: None,
+        505: None,
+        507: None,
+        548: None,
+        592: None,
+        604: None,
+        691: None,
+        692: None,
+        1006: None,
+        1007: None,
+        1097: None,
+    },
 }
