+++ tests/integration/test_data/open3e_device_config/Open3Edatapoints_6a1.py
@@ -3,15 +3,14 @@ from open3e.Open3Ecodecs import *
 
 dataIdentifiers = {
     "name": "EMCUMASTER",
-    "dids" :
-    {
-        256 : None,
-        262 : None,
-        382 : None,
-        505 : None,
-        507 : None,
-        592 : None,
-        604 : None,
-        2214 : None,
-    }
+    "dids": {
+        256: None,
+        262: None,
+        382: None,
+        505: None,
+        507: None,
+        592: None,
+        604: None,
+        2214: None,
+    },
 }
